refactor(analyzer): Move `Graph.toPackageReferenceForest()`
Prepare for making the graph operate on (name, version) tuples.
